Fix a leak of the handler data
Previously the `box callback` expression was being leaked because no on ever got
around to freeing it. This alters the server to store it by value so its dropped
when the server exits.

This prevents use-after-free by shutting down the server in the destructor,
before the data is dropped, so it's know that the data is longer in use by C.